service/requirements.txt
# Core gRPC and service dependencies
grpcio>=1.50.0
grpcio-tools>=1.50.0
grpcio-health-checking>=1.50.0
protobuf>=4.21.0

# Serialization and data handling
cloudpickle>=2.2.0
numpy>=1.21.0

# WebRTC support
aiortc>=1.13.0
aiohttp>=3.10.0
av>=14.0.0

# Security and sandboxing
psutil>=5.9.0

# Configuration and logging
pyyaml>=6.0
structlog>=22.0.0

# Async support
asyncio-mqtt>=0.11.0
aiohttp-cors>=0.7.0

# Monitoring and metrics
prometheus-client>=0.15.0

# SDK dependencies are now installed via the remotemedia package
# which is installed directly from source in the Dockerfile

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-grpc>=0.8.0 
librosa>=0.11.0